fix: Fixed backend uuid url and, implemented community pages in url, also fixed messaging not working.
