+++ web/src/routes/_authorized/discover.tsx
@@ -14,6 +14,7 @@ import {
     recentlyAddedMentorsQueryOptions,
 } from '@/lib/queries/DiscoverQueries.ts'
 import { TrendingUp, Sparkles } from 'lucide-react'
+import { useSendMessageToUser } from '@/lib/queries/MessagingQueries.ts'
 
 const PAGE_SIZE = 6
 const DISCOVER_SECTION_CONTAINER_CLASS =
@@ -33,9 +34,10 @@ interface MentorRowProps {
     icon: React.ReactNode
     profiles: import('@/lib/queries/DiscoverQueries.ts').PublicMentorProfile[]
     onViewProfile: (username: string) => void
+    onSendMessage: (username: string) => void
 }
 
-function MentorRow({ title, subtitle, icon, profiles, onViewProfile }: MentorRowProps) {
+function MentorRow({ title, subtitle, icon, profiles, onViewProfile, onSendMessage }: MentorRowProps) {
     return (
         <div className={`${DISCOVER_SECTION_CONTAINER_CLASS} flex flex-col gap-4`}>
             <div className="flex items-end justify-between">
@@ -57,7 +59,7 @@ function MentorRow({ title, subtitle, icon, profiles, onViewProfile }: MentorRow
                         <ProfileCard
                             profile={profile}
                             onViewProfile={onViewProfile}
-                            onSendMessage={() => {}}
+                            onSendMessage={onSendMessage}
                             className="h-full"
                         />
                     </div>
@@ -73,6 +75,7 @@ function MentorRow({ title, subtitle, icon, profiles, onViewProfile }: MentorRow
 
 export function DiscoverPage() {
     const navigate = useNavigate()
+    const handleSendMessage = useSendMessageToUser()
     const [query, setQuery] = useState('')
     const [selectedSkills, setSelectedSkills] = useState<Set<string>>(new Set())
 
@@ -155,6 +158,7 @@ export function DiscoverPage() {
                             icon={<TrendingUp className="h-5 w-5 text-accent" />}
                             profiles={popularMentors}
                             onViewProfile={handleViewProfile}
+                            onSendMessage={handleSendMessage}
                         />
                     )}
 
@@ -165,6 +169,7 @@ export function DiscoverPage() {
                             icon={<Sparkles className="h-5 w-5 text-amber-500" />}
                             profiles={recentMentors}
                             onViewProfile={handleViewProfile}
+                            onSendMessage={handleSendMessage}
                         />
                     )}
 
@@ -201,7 +206,7 @@ export function DiscoverPage() {
                                 key={profile.id}
                                 profile={profile}
                                 onViewProfile={handleViewProfile}
-                                onSendMessage={() => {}}
+                                onSendMessage={handleSendMessage}
                                 className="h-full"
                             />
                         ))}
